% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aphylo2-terms}
\alias{aphylo2-terms}
\alias{term_gains}
\alias{term_loss}
\alias{term_cogain}
\alias{term_neofun}
\alias{term_subfun}
\alias{term_maxfuns}
\alias{term_overall_changes}
\alias{term_kgains}
\alias{term_neofun_a2b}
\alias{term_genes_changing}
\title{Evolutionary terms}
\usage{
term_gains(p, funs, duplication = TRUE)

term_loss(p, funs, duplication = TRUE)

term_cogain(p, a, b)

term_neofun(p, a, b)

term_subfun(p, a, b)

term_maxfuns(p, lb, ub, duplication = TRUE)

term_overall_changes(p, duplication = TRUE)

term_kgains(p, funs, k = 1L, duplication = TRUE)

term_neofun_a2b(p, a, b, duplication = TRUE)

term_genes_changing(p, duplication = TRUE)
}
\description{
Evolutionary terms
}
