% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo2_mcmc.R
\name{aphylo2_mcmc}
\alias{aphylo2_mcmc}
\title{MCMC for aphylo2}
\usage{
aphylo2_mcmc(amodel, initial = rep(0, nterms(amodel)), ...)
}
\arguments{
\item{amodel}{an object of class \linkS4class{aphylo2}}

\item{initial}{Vector of initial parameters}

\item{...}{Passed to \link[fmcmc:MCMC]{fmcmc::MCMC}}
}
\description{
MCMC for aphylo2
}
