% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo2_mle.R
\name{aphylo2_mle}
\alias{aphylo2_mle}
\title{MLE for aphylo2}
\usage{
aphylo2_mle(
  amodel,
  initial = rep(0, nterms(amodel)),
  control = list(maxit = 1000L, fnscale = -1),
  ...
)
}
\arguments{
\item{amodel}{an object of class \linkS4class{aphylo2}}

\item{initial}{Vector of initial parameters}

\item{...}{Passed to \link{stats:optim}}
}
\description{
MLE for aphylo2
}
