% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geese-class}
\alias{geese-class}
\alias{new_geese}
\alias{geese}
\title{GEne Evolutionary model using SufficiEncy (GEESE)}
\usage{
new_geese(annotations, geneid, parent, duplication)
}
\arguments{
\item{annotations}{Vector of integer vectors with annotations.}

\item{geneid}{integer vector with gene ids.}

\item{parent}{integer vector with parent gene id.}

\item{duplication}{logical vector indicating the type of events.}

\item{p}{An object of class \code{geese}.}
}
\description{
GEne Evolutionary model using SufficiEncy (GEESE)
}
