% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/methods.R
\name{geese-common}
\alias{geese-common}
\alias{init_model}
\alias{flock-common}
\alias{nterms}
\alias{nnodes}
\alias{ntrees}
\alias{nleafs}
\alias{likelihood}
\alias{get_probabilities}
\alias{get_sequence}
\alias{set_seed}
\alias{sim_geese}
\alias{observed_counts}
\alias{print_observed_counts}
\alias{support_size}
\alias{parse_polytomies}
\alias{nfuns}
\alias{names.geese}
\alias{predict_geese}
\alias{predict_flock}
\alias{names.flock}
\alias{print.geese}
\alias{print.flock}
\title{Common functions for \code{geese} and \code{flock}.}
\usage{
init_model(p, verb = TRUE)

nterms(p)

nnodes(p)

ntrees(p)

nleafs(p)

likelihood(p, par, as_log = FALSE, trunc_seq = TRUE)

get_probabilities(p)

get_sequence(p, reduced_sequence = TRUE)

set_seed(p, s)

sim_geese(p, par, seed = -1L)

observed_counts(p)

print_observed_counts(p)

support_size(p)

parse_polytomies(p)

nfuns(p)

\method{names}{geese}(p)

predict_geese(
  p,
  par,
  leave_one_out = FALSE,
  use_reduced_sequence = TRUE,
  only_annotated = FALSE
)

predict_flock(
  p,
  par,
  leave_one_out = FALSE,
  use_reduced_sequence = TRUE,
  only_annotated = FALSE
)

\method{names}{flock}(x)

\method{print}{geese}(x, ...)

\method{print}{flock}(x, ...)
}
\arguments{
\item{p}{An object of class \link{geese} or \link{flock}.}

\item{verb}{Logical scalar. When \code{TRUE}, it will print a progress bar during
the initialization of the process.}

\item{par}{Numeric vector of length \code{nterms()}.}

\item{trunc_seq}{When \code{TRUE} uses the truncated pruning sequence (see details).}
}
\value{
\code{nterms} returns the number of terms included in the model. This
is different from the number of parameters as the later includes the number
of functions in the data.

\code{nnodes} returns the number of nodes in the data, this includes
internal nodes. If \code{p} is a \link{flock}, then it will return a vector of length
\code{ntrees()}.

\code{ntrees} returns the number of trees in the model. For a geese object
this will be equal to one.
}
\description{
Common functions for \code{geese} and \code{flock}.
}
\details{
\code{init_model} initializes the model. This triggers the calculation
of the support using the vector of terms included. Initializing a model can
only be done once.

Using the truncated pruning sequence (\code{trunc_seq = TRUE}) involves
traversing the trees throught the induced subtree. This is relevant in the case
that not all the leafs are annotated.
}
