% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geese-terms}
\alias{geese-terms}
\alias{rule_limit_changes}
\alias{term_gains}
\alias{term_loss}
\alias{term_cogain}
\alias{term_neofun}
\alias{term_subfun}
\alias{term_maxfuns}
\alias{term_overall_changes}
\alias{term_kgains}
\alias{term_neofun_a2b}
\alias{term_genes_changing}
\alias{term_coopt}
\title{Evolutionary terms}
\usage{
rule_limit_changes(p, term_pos, lb, ub, duplication = TRUE)

term_gains(p, funs, duplication = TRUE)

term_loss(p, funs, duplication = TRUE)

term_cogain(p, a, b)

term_neofun(p, a, b)

term_subfun(p, a, b)

term_maxfuns(p, lb, ub, duplication = TRUE)

term_overall_changes(p, duplication = TRUE)

term_kgains(p, funs, k = 1L, duplication = TRUE)

term_neofun_a2b(p, a, b, duplication = TRUE)

term_genes_changing(p, duplication = TRUE)

term_coopt(p, a, b, duplication = TRUE)
}
\arguments{
\item{p}{An object of class \link{geese} or \link{flock}.}

\item{duplication}{When \code{TRUE} indicates that this term is only valid for
duplication events.}

\item{funs}{Vector of function indices (starting from zero).}

\item{a, b}{Indices of functions (starting from zero)}
}
\description{
Model terms for both \link{geese} and \link{flock} objects.

Model terms for both \link{geese} and \link{flock} objects.
}
\details{
In the case of \code{term_neofun_a2b}, \code{a} represents the origin function
from which \code{b} is originated.
}
