% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geese_mcmc.R
\name{geese_mcmc}
\alias{geese_mcmc}
\title{MCMC for geese}
\usage{
geese_mcmc(
  amodel,
  initial = rep(0, nterms(amodel)),
  prior = function(p) {     stats::dlogis(p, log = TRUE) },
  ...
)
}
\arguments{
\item{amodel}{an object of class \linkS4class{geese}}

\item{initial}{Vector of initial parameters}

\item{...}{Passed to \link[fmcmc:MCMC]{fmcmc::MCMC}}
}
\description{
MCMC for geese
}
