% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geese_mle.R
\name{geese_mle}
\alias{geese_mle}
\title{MLE for geese}
\usage{
geese_mle(
  amodel,
  initial = rep(0, nterms(amodel)),
  control = list(maxit = 1000L, fnscale = -1),
  ...
)
}
\arguments{
\item{amodel}{an object of class \linkS4class{geese}}

\item{initial}{Vector of initial parameters}

\item{...}{Passed to \link{stats:optim}}
}
\description{
MLE for geese
}
