% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transition_prob}
\alias{transition_prob}
\alias{conditional_prob}
\title{Compute the transition probability}
\usage{
transition_prob(p, params, duplication, state, array, as_log = FALSE)

conditional_prob(p, params, duplication, state, array, i, j, as_log = FALSE)
}
\arguments{
\item{p}{An object of class \code{geese} or \code{flock}.}

\item{params}{A vector of model parameters.}

\item{duplication}{logical scalar. Type of evolutionary event.}

\item{state}{logical vector. State of the parent.}

\item{array}{matrix indicating the state of the offspring (rows = function, cols = offspring).}

\item{as_log}{logical scalar. When \code{TRUE} returns the log.}

\item{i, j}{Locations (index from zero) of the cell to compute.}
}
\description{
Compute the transition probability
}
\details{
\code{conditional_prob} computes the so-called "Gibbs sampling"
probability, in which the likelihood of observing Y(i,j) = 1 conditional
on the rest of the data is computed.
}
