% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aphylo2-class}
\alias{aphylo2-class}
\alias{new_model}
\alias{init}
\alias{likelihood}
\alias{get_probabilities}
\alias{get_sequence}
\alias{sim_aphylo2}
\title{Aphylo2 model}
\usage{
new_model(annotations, geneid, parent)

init(p)

likelihood(p, par)

get_probabilities(p)

get_sequence(p)

sim_aphylo2(p, par)
}
\arguments{
\item{annotations}{Vector of integer vectors with annotations.}

\item{geneid}{integer vector with gene ids}

\item{parent}{integer vector with parent gene id}
}
\description{
Aphylo2 model
}
