% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aphylo2-terms}
\alias{aphylo2-terms}
\alias{term_gains}
\alias{term_loss}
\alias{term_cogain}
\alias{term_neofun}
\alias{term_subfun}
\alias{term_maxfuns}
\title{Evolutionary terms}
\usage{
term_gains(p, funs, duplication = TRUE)

term_loss(p, funs, duplication = TRUE)

term_cogain(p, a, b)

term_neofun(p, a, b)

term_subfun(p, a, b)

term_maxfuns(p, lb, ub)
}
\description{
Evolutionary terms
}
