% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geese_mle.R
\name{geese_mle}
\alias{geese_mle}
\title{MLE for geese}
\usage{
geese_mle(
  amodel,
  initial = rep(0, nterms(amodel)),
  control = list(),
  offset_list = NULL,
  ncores = 1L,
  ...
)
}
\arguments{
\item{amodel}{an object of class \linkS4class{geese}}

\item{initial}{Vector of initial parameters}

\item{offset_list}{Integer vector with the indices of the parameters
that will set set to be fixed.}

\item{...}{Passed to \code{\link[stats:optim]{stats::optim()}}}
}
\description{
MLE for geese
}
